+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Unreleased
 
 - Added `request_user_attention` method to `Window`.
-- **Breaking:** On macOS, removed `WindowExt::request_user_attention`.
+- **Breaking:** On macOS, removed `WindowExt::request_user_attention`, use `Window::request_user_attention`.  
+- **Breaking:** On X11, removed `WindowExt::set_urgent`, use `Window::request_user_attention`. 
 - On Android, calling `WindowEvent::Focused` now works properly instead of always returning false. 
 - On Windows, fix alt-tab behaviour by removing borderless fullscreen "always on top" flag.
 - On Windows, fix bug preventing windows with transparency enabled from having fully-opaque regions.